requirements.txt
numpy
torch
matplotlib
pillow
PyYAML
tqdm
opencv-python
scikit-learn